Move attribute set to root crate
