+++ opentelemetry-sdk/src/attributes/mod.rs
@@ -1,3 +1,2 @@
 mod set;
 
-pub use set::AttributeSet;
