+++ opentelemetry-sdk/src/metrics/data/mod.rs
@@ -4,7 +4,8 @@ use std::{any, borrow::Cow, fmt, time::SystemTime};
 
 use opentelemetry::{metrics::Unit, KeyValue};
 
-use crate::{attributes::AttributeSet, instrumentation::Scope, Resource};
+use opentelemetry::attributes::AttributeSet;
+use crate::{instrumentation::Scope, Resource};
 
 pub use self::temporality::Temporality;
