+++ opentelemetry-sdk/src/metrics/internal/exponential_histogram.rs
@@ -6,11 +6,10 @@ use std::{
 };
 
 use once_cell::sync::Lazy;
-use opentelemetry::metrics::MetricsError;
+use opentelemetry::{attributes::AttributeSet, metrics::MetricsError};
 
 use crate::{
     metrics::data::{self, Aggregation, Temporality},
-    AttributeSet,
 };
 
 use super::Number;
