+++ opentelemetry-sdk/src/metrics/internal/last_value.rs
@@ -4,8 +4,8 @@ use std::{
     time::SystemTime,
 };
 
-use crate::{attributes::AttributeSet, metrics::data::DataPoint};
-use opentelemetry::{global, metrics::MetricsError};
+use crate::{metrics::data::DataPoint};
+use opentelemetry::{attributes::AttributeSet, global, metrics::MetricsError};
 
 use super::{
     aggregate::{is_under_cardinality_limit, STREAM_OVERFLOW_ATTRIBUTE_SET},
