+++ opentelemetry-sdk/src/metrics/internal/sum.rs
@@ -4,9 +4,8 @@ use std::{
     time::SystemTime,
 };
 
-use crate::attributes::AttributeSet;
 use crate::metrics::data::{self, Aggregation, DataPoint, Temporality};
-use opentelemetry::{global, metrics::MetricsError};
+use opentelemetry::{attributes::AttributeSet, global, metrics::MetricsError};
 
 use super::{
     aggregate::{is_under_cardinality_limit, STREAM_OVERFLOW_ATTRIBUTE_SET},
