+++ opentelemetry-stdout/src/logs/transform.rs
@@ -4,7 +4,7 @@ use crate::common::{
     as_human_readable, as_opt_human_readable, as_opt_unix_nano, as_unix_nano, KeyValue, Resource,
     Scope, Value,
 };
-use opentelemetry_sdk::AttributeSet;
+use opentelemetry::attributes::AttributeSet;
 use serde::Serialize;
 
 /// Transformed logs data that can be serialized.
