+++ opentelemetry/Cargo.toml
@@ -24,6 +24,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 futures-core = "0.3"
 futures-sink = "0.3"
 once_cell = "1.12.0"
+ordered-float = "4.0"
 pin-project-lite = { version = "0.2", optional = true }
 thiserror = "1.0.7"
 urlencoding = "2.1.2"
