+++ opentelemetry/src/attributes/mod.rs
@@ -1,0 +1,3 @@
+mod set;
+
+pub use set::AttributeSet;
